Overhaul for new site
It has been quite a while since a new version of this site has taken shape!

This overhaul is for deployment into the Netlify's free hosting tier. This is a single
page, static site which shows the latest scheduled event. The event is powered via two
requests:

* meetup.com's atom feed for the latest event
* specific event on meetup.com parsed for opengraph data

This is done using Netlify's serverless function option. Cache busting and a client side
build step, for minification, etc, should be added. If you're looking to contribute, reach
out or open a PR and we're happy to assist!